refactor where to fetch example collection, fix tests
