+++ src/models/Collection.test.ts
@@ -4,6 +4,7 @@ import User from './User'
 import { CollectionModel } from './Collection'
 import { MySpaceModel } from './MySpace'
 import { CollectionInput, CollectionRecords } from 'types'
+import { EXAMPLE_COLLECTION } from '__fixtures__/newPortalUser'
 
 let connection: typeof MongoClient
 let db: typeof Db
@@ -141,7 +142,7 @@ describe('Collection Model', () => {
 
     // Create test user
     testUserId = 'testUserId'
-    await User.createOne(testUserId, { db })
+    await User.createOne(testUserId, EXAMPLE_COLLECTION, { db })
     const testUser = await User.findOne(testUserId, { db })
     exampleCollectionId = testUser.mySpace[0]._id
   })
