+++ src/models/User.test.ts
@@ -1,49 +1,10 @@
 import { Db, MongoClient, ObjectId } from 'mongodb'
 import User from './User'
-import type { Collection } from 'types/index'
-import { WIDGET_TYPES } from 'constants/index'
+import { EXAMPLE_COLLECTION } from '../__fixtures__/newPortalUser'
 
 let connection: typeof MongoClient
 let db: typeof Db
 
-const EXAMPLE_COLLECTION: Collection = {
-  _id: ObjectId(),
-  cmsId: 'ckwz3u58s1835ql974leo1yll',
-  title: 'Example Collection',
-  type: WIDGET_TYPES.COLLECTION,
-  bookmarks: [
-    {
-      _id: ObjectId(),
-      cmsId: 'cktd7c0d30190w597qoftevq1',
-      url: 'https://afpcsecure.us.af.mil/',
-      label: 'vMPF',
-    },
-    {
-      _id: ObjectId(),
-      cmsId: 'cktd7ettn0457w597p7ja4uye',
-      url: 'https://leave.af.mil/profile',
-      label: 'LeaveWeb',
-    },
-    {
-      _id: ObjectId(),
-      cmsId: 'cktd7hjz30636w5977vu4la4c',
-      url: 'https://mypay.dfas.mil/#/',
-      label: 'MyPay',
-    },
-    {
-      _id: ObjectId(),
-      cmsId: 'ckwz3tphw1763ql97pia1zkvc',
-      url: 'https://webmail.apps.mil/',
-      label: 'Webmail',
-    },
-    {
-      _id: ObjectId(),
-      cmsId: 'ckwz3u4461813ql970wkd254m',
-      url: 'https://www.e-publishing.af.mil/',
-      label: 'e-Publications',
-    },
-  ],
-}
 describe('User model', () => {
   beforeAll(async () => {
     connection = await MongoClient.connect(process.env.MONGO_URL, {
@@ -67,7 +28,7 @@ describe('User model', () => {
       mySpace: [EXAMPLE_COLLECTION],
     }
 
-    await User.createOne('testUserId', { db })
+    await User.createOne('testUserId', EXAMPLE_COLLECTION, { db })
 
     const insertedUser = await User.findOne('testUserId', { db })
